+++ modules/renderer/src/com/genir/renderer/bridge/commands/GL11.java
@@ -904,7 +904,14 @@ public class GL11 {
      */
     public static void glFlush() {
         // Don't do anything. glFlush and glFinish are mostly
-        // redundant when Display update is being called.
+        // redundant when Display.update() is being called.
+
+        // BoxUtil 1.5.4 requires glFlush() in parallel contexts to render
+        // certain entities correctly. However, with parallel GL calls
+        // synchronized by Fast Rendering, glFlush() calls from modded code
+        // are redundant. Fast Rendering performs the required flush at
+        // synchronization points in the GL method executor. This reduces
+        // the number of flushes and CPU load while preserving correctness.
     }
 
     public static void glFinish() {