+++ changelog.md
@@ -1,3 +1,8 @@
+v0.8.4rc1
+
+- Improved rendering stability and performance when using BoxUtil.
+- Fixed missing entities and image flicker that could occur on certain low-end GPUs when using BoxUtil. 
+
 v0.8.3
  
 - Fixed a NullPointerException when loading PMMM with PMMMVE. Reported by AlfadorZero, hyperspace turtle.
@@ -14,7 +19,7 @@ v0.8.1
 - Minor rendering performance optimizations.
 - Added `fr.noterminal.bat`, which starts the game faster. A terminal window still appears during startup, but it can be closed safely.
 - Optimized game startup time.
-- 
+
 v0.8.0
 
 - Ported the bytecode modification pipeline from a custom system class loader to a Java agent.